refactor: put  reacquireInputTrack to local audio stream
